// src/app/core/interfaces/interface.pedidos.ts
export interface Pedidos {

    idpedido?: number;
    idpedidoant?: string;
    idsolicitante?: number;
    idpersonalant?: string;
    fecha?: string;
    destino?: string;
    despachado?: string;
    departamento?: string;
    observacion?: string;
    preciopedido?: number;
    hora?: string;
    usuarioreg?: string;
    fechareg?: string;
    usuariomod?: string;
    fechamod?: string;
    
}